still working on...

+++ view.js
@@ -8,21 +8,64 @@ getGamesList(function(arrayOfGames){
 function createDomElement(gameObj){
     var container1 = document.querySelector('.container');
     const gameELement = document.createElement("div");
-    gameELement.innerHTML = `<h1>${gameObj.title}</h1> 
+    gameELement.innerHTML = `<div id="${gameObj._id}">
+                        <h1>${gameObj.title}</h1> 
                         <img src="${gameObj.imageUrl}" />
                         <p>${gameObj.description}</p> 
-                        <button class="delete-btn" id="${gameObj._id}">Delete Game</button>
-                        <button class="update-btn" id="${gameObj._id}">Edit Game</button>`;    
+                        <button class="delete-btn">Delete Game</button>
+                        <button class="update-btn">Edit Game</button></div>`;    
 
     container1.appendChild(gameELement);
 
     document.getElementById(`${gameObj._id}`).addEventListener("click", function(event){
-        deleteGame(event.target.getAttribute("id"), function(apiResponse){
-            console.log(apiResponse);
-            removeDeletedElementFromDOM(event.target.parentElement);
-        })
+        if(event.target.classList.contains("delete-btn")){
+            console.log(event.target);
+            deleteGame(event.target.getAttribute("id"), function(apiResponse){
+                console.log(apiResponse);
+                removeDeletedElementFromDOM(event.target.parentElement);
+            })
+        }else if(event.target.classList.contains("update-btn")){
+            console.log("ce e aici?")
+            // var toUpdateObject ={
+            //     title:event.target.parentElement.querySelector("h1").innerHTML,
+            //     description: ,
+            //     imageUrl:
+            // }
+            var toUpdateObject = {
+                title : event.target.parentElement.querySelector("h1").innerHTML,
+                description : event.target.parentElement.querySelector("p").innerHTML,
+                imageUrl:"https://miro.medium.com/max/1200/1*mk1-6aYaf_Bes1E3Imhc0A.jpeg"
+            };
+
+            createUpdateDivElement( toUpdateObject, event.target.parentElement);
+            
+
+        }
+        
     });
 }
+
+function createUpdateDivElement(updateGameObj, gameContainer){
+    console.log(updateGameObj);
+    const updateElement = document.createElement("div");
+    updateElement.classList.add("updateForm")
+    updateElement.setAttribute("class","updateForm")
+    updateElement.innerHTML = 
+        `<label for="gameTitle">Your Title *</label>
+        <input type="text" value="${updateGameObj.title}" name="gameTitle" id="updateGameTitle" />
+
+        <label for="gameDescription">Your Description</label>
+        <textarea name="gameDescription" value = "${updateGameObj.description}" id="updateGameDescription"></textarea>
+
+        <label for="gameImageUrl">Put a new URL image*</label>
+        <input type="text" value = "${updateGameObj.imageUrl}" name="updateGameImageUrl" id="gameImageUrl" />
+
+        <button class="updateBtn">Save Changes</button>
+        <button class="cancelBtn">Cancel</button>`
+    gameContainer.appendChild(updateElement)
+}
+
+
 
 function removeDeletedElementFromDOM(domElement){
     domElement.remove();
@@ -87,3 +130,13 @@ document.querySelector(".submitBtn").addEventListener("click", function(event){
         createGameRequest(urlencoded, createDomElement);
     }
 })
+
+
+// const updateGameTitle = document.getElementById("updateGameTitle");
+// const updateGameDescription = document.getElementById("updateGameDescription");
+// const updateGameImageUrl = document.getElementById("updateGameImageUrl");
+
+// var urlencoded2 = new URLSearchParams();
+// urlencoded2.append("title", updateGameTitle.value);
+// urlencoded2.append("description", updateGameDescription.value);
+// urlencoded2.append("imageUrl", updateGameImageUrl.value);
